requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
sqlparse==0.5.0
